requirements.txt
Pillow==9.3.0
botocore==1.29.21
boto3==1.26.21
uvicorn==0.19.0
fastapi==0.86.0
pydantic==1.10.2
python-multipart==0.0.5